+++ internal/scaffold/helm/go_mod.go
@@ -1,0 +1,101 @@
+// Copyright 2019 The Operator-SDK Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package helm
+
+import (
+	"fmt"
+
+	"github.com/operator-framework/operator-sdk/internal/scaffold/input"
+	"github.com/operator-framework/operator-sdk/internal/scaffold/internal/deps"
+)
+
+const GoModFile = "go.mod"
+
+// GoMod - the go.mod file for a Helm hybrid operator.
+type GoMod struct {
+	input.Input
+}
+
+func (s *GoMod) GetInput() (input.Input, error) {
+	if s.Path == "" {
+		s.Path = GoModFile
+	}
+	s.TemplateBody = goModTmpl
+	return s.Input, nil
+}
+
+const goModTmpl = `module {{ .Repo }}
+
+go 1.13
+
+require (
+	github.com/Azure/go-ansiterm v0.0.0-20170929234023-d6e3b3328b78 // indirect
+	github.com/coreos/pkg v0.0.0-20180928190104-399ea9e2e55f // indirect
+	github.com/coreos/prometheus-operator v0.31.1 // indirect
+	github.com/docker/docker v0.0.0-20180612054059-a9fbbdc8dd87 // indirect
+	github.com/emicklei/go-restful v2.9.3+incompatible // indirect
+	github.com/go-openapi/spec v0.19.0 // indirect
+	github.com/google/btree v1.0.0 // indirect
+	github.com/gorilla/websocket v1.4.0 // indirect
+	github.com/gotestyourself/gotestyourself v2.2.0+incompatible // indirect
+	github.com/gregjones/httpcache v0.0.0-20190212212710-3befbb6ad0cc // indirect
+	github.com/grpc-ecosystem/go-grpc-prometheus v1.2.0 // indirect
+	github.com/imdario/mergo v0.3.7 // indirect
+	github.com/jonboulle/clockwork v0.1.0 // indirect
+	github.com/liggitt/tabwriter v0.0.0-20181228230101-89fcab3d43de // indirect
+	github.com/operator-framework/operator-sdk master
+	github.com/spf13/pflag v1.0.3
+	github.com/ugorji/go/codec v0.0.0-20190320090025-2dc34c0b8780 // indirect
+	golang.org/x/time v0.0.0-20190308202827-9d24e82272b4 // indirect
+	google.golang.org/appengine v1.5.0 // indirect
+	gotest.tools v2.2.0+incompatible // indirect
+	k8s.io/apiserver v0.0.0-20181213151703-3ccfe8365421 // indirect
+	k8s.io/client-go v11.0.1-0.20190409021438-1a26190bd76a+incompatible
+	k8s.io/kube-openapi v0.0.0-20190603182131-db7b694dc208 // indirect
+	sigs.k8s.io/controller-runtime v0.2.0
+)
+
+// Pinned to kubernetes-1.14.1
+replace (
+	k8s.io/api => k8s.io/api kubernetes-1.14.1
+	k8s.io/apiextensions-apiserver => k8s.io/apiextensions-apiserver kubernetes-1.14.1
+	k8s.io/apimachinery => k8s.io/apimachinery kubernetes-1.14.1
+	k8s.io/apiserver => k8s.io/apiserver kubernetes-1.14.1
+	k8s.io/cli-runtime => k8s.io/cli-runtime kubernetes-1.14.1
+	k8s.io/client-go => k8s.io/client-go kubernetes-1.14.1
+	k8s.io/cloud-provider => k8s.io/cloud-provider kubernetes-1.14.1
+	k8s.io/kubernetes => k8s.io/kubernetes v1.14.1
+)
+
+replace (
+	// Indirect operator-sdk dependencies use git.apache.org, which is frequently
+	// down. The github mirror should be used instead.
+	// Locking to a specific version (from 'go mod graph'):
+	git.apache.org/thrift.git => github.com/apache/thrift v0.0.0-20180902110319-2566ecd5d999
+	github.com/coreos/prometheus-operator => github.com/coreos/prometheus-operator v0.31.1
+	// Pinned to v2.10.0 (kubernetes-1.14.1) so https://proxy.golang.org can
+	// resolve it correctly.
+	github.com/prometheus/prometheus => github.com/prometheus/prometheus d20e84d0fb64aff2f62a977adc8cfb656da4e286
+)
+`
+
+func PrintGoMod() error {
+	b, err := deps.ExecGoModTmpl(goModTmpl)
+	if err != nil {
+		return err
+	}
+	fmt.Print(string(b))
+	return nil
+}